Move config endpoints to an independent file.
